requirements.txt
pyapi-gitlab==7.8.5
requests==2.10.0
wheel==0.24.0
